+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/) and [Keep A Changelog's Format](http://keepachangelog.com/).
 
+## [2.31.0] 2018-02-28
+
+### Added
+- Footer Component
+
 ## [2.30.0] 2018-02-07
 
 ### Added